--- projetos/Semantic-Compression/semantic-compression/modeling/pspnet/config/cityscapes_colors.txt
@@ -1,19 +1,0 @@
-128 64 128
-244 35 232
-70 70 70
-102 102 156
-190 153 153
-153 153 153
-250 170 30
-220 220 0
-107 142 35
-152 251 152
-70 130 180
-220 20 60
-255 0 0
-0 0 142
-0 0 70
-0 60 100
-0 80 100
-0 0 230
-119 11 32
