--- projetos/Semantic-Compression/semantic-compression/modeling/pspnet/config/cityscapes_names.txt
@@ -1,19 +1,0 @@
-road
-sidewalk
-building
-wall
-fence
-pole
-traffic light
-traffic sign
-vegetation
-terrain
-sky
-person
-rider
-car
-truck
-bus
-train
-motorcycle
-bicycle
