--- projetos/Semantic-Compression/semantic-compression/modeling/pspnet/config/cityscapes_pspnet101.yaml
@@ -1,58 +1,0 @@
-DATA:
-  data_root: dataset/cityscapes
-  train_list: dataset/cityscapes/list/fine_train.txt
-  val_list: dataset/cityscapes/list/fine_val.txt
-  classes: 19
-
-TRAIN:
-  arch: psp
-  layers: 101
-  sync_bn: True  # adopt syncbn or not
-  train_h: 713
-  train_w: 713
-  scale_min: 0.5  # minimum random scale
-  scale_max: 2.0  # maximum random scale
-  rotate_min: -10  # minimum random rotate
-  rotate_max: 10  # maximum random rotate
-  zoom_factor: 8  # zoom factor for final prediction during training, be in [1, 2, 4, 8]
-  ignore_label: 255
-  aux_weight: 0.4
-  train_gpu: [0, 1, 2, 3, 4, 5, 6, 7]
-  workers: 16  # data loader workers
-  batch_size: 16  # batch size for training
-  batch_size_val: 8  # batch size for validation during training, memory and speed tradeoff
-  base_lr: 0.01
-  epochs: 200
-  start_epoch: 0
-  power: 0.9
-  momentum: 0.9
-  weight_decay: 0.0001
-  manual_seed:
-  print_freq: 10
-  save_freq: 1
-  save_path: exp/cityscapes/pspnet101/model
-  weight:  # path to initial weight (default: none)
-  resume:  # path to latest checkpoint (default: none)
-  evaluate: False  # evaluate on validation set, extra gpu memory needed and small batch_size_val is recommend
-Distributed:
-  dist_url: tcp://127.0.0.1:6789
-  dist_backend: 'nccl'
-  multiprocessing_distributed: True
-  world_size: 1
-  rank: 0
-
-TEST:
-  test_list: dataset/cityscapes/list/fine_val.txt
-  split: val  # split in [train, val and test]
-  base_size: 2048  # based size for scaling
-  test_h: 713
-  test_w: 713
-  scales: [1.0]  # evaluation scales, ms as [0.5, 0.75, 1.0, 1.25, 1.5, 1.75]
-  has_prediction: False  # has prediction already or not
-  index_start: 0  # evaluation start index in list
-  index_step: 0  # evaluation step index in list, 0 means to end
-  test_gpu: [0]
-  model_path: config/train_epoch_200.pth  # evaluation model path
-  save_folder: exp/cityscapes/pspnet101/result/epoch_200/val/ss  # results save folder
-  colors_path: config/cityscapes_colors.txt  # path of dataset colors
-  names_path: config/cityscapes_names.txt  # path of dataset category names
