--- projetos/Semantic-Compression/semantic-compression/modeling/pspnet/util/util.py
@@ -1,169 +1,0 @@
-import os
-import numpy as np
-from PIL import Image
-
-import torch
-from torch import nn
-import torch.nn.init as initer
-
-
-class AverageMeter(object):
-    """Computes and stores the average and current value"""
-    def __init__(self):
-        self.reset()
-
-    def reset(self):
-        self.val = 0
-        self.avg = 0
-        self.sum = 0
-        self.count = 0
-
-    def update(self, val, n=1):
-        self.val = val
-        self.sum += val * n
-        self.count += n
-        self.avg = self.sum / self.count
-
-
-def step_learning_rate(base_lr, epoch, step_epoch, multiplier=0.1):
-    """Sets the learning rate to the base LR decayed by 10 every step epochs"""
-    lr = base_lr * (multiplier ** (epoch // step_epoch))
-    return lr
-
-
-def poly_learning_rate(base_lr, curr_iter, max_iter, power=0.9):
-    """poly learning rate policy"""
-    lr = base_lr * (1 - float(curr_iter) / max_iter) ** power
-    return lr
-
-
-def intersectionAndUnion(output, target, K, ignore_index=255):
-    # 'K' classes, output and target sizes are N or N * L or N * H * W, each value in range 0 to K - 1.
-    assert (output.ndim in [1, 2, 3])
-    assert output.shape == target.shape
-    output = output.reshape(output.size).copy()
-    target = target.reshape(target.size)
-    output[np.where(target == ignore_index)[0]] = ignore_index
-    intersection = output[np.where(output == target)[0]]
-    area_intersection, _ = np.histogram(intersection, bins=np.arange(K+1))
-    area_output, _ = np.histogram(output, bins=np.arange(K+1))
-    area_target, _ = np.histogram(target, bins=np.arange(K+1))
-    area_union = area_output + area_target - area_intersection
-    return area_intersection, area_union, area_target
-
-
-def intersectionAndUnionGPU(output, target, K, ignore_index=255):
-    # 'K' classes, output and target sizes are N or N * L or N * H * W, each value in range 0 to K - 1.
-    assert (output.dim() in [1, 2, 3])
-    assert output.shape == target.shape
-    output = output.view(-1)
-    target = target.view(-1)
-    output[target == ignore_index] = ignore_index
-    intersection = output[output == target]
-    area_intersection = torch.histc(intersection, bins=K, min=0, max=K-1)
-    area_output = torch.histc(output, bins=K, min=0, max=K-1)
-    area_target = torch.histc(target, bins=K, min=0, max=K-1)
-    area_union = area_output + area_target - area_intersection
-    return area_intersection, area_union, area_target
-
-
-def check_mkdir(dir_name):
-    if not os.path.exists(dir_name):
-        os.mkdir(dir_name)
-
-
-def check_makedirs(dir_name):
-    if not os.path.exists(dir_name):
-        os.makedirs(dir_name)
-
-
-def init_weights(model, conv='kaiming', batchnorm='normal', linear='kaiming', lstm='kaiming'):
-    """
-    :param model: Pytorch Model which is nn.Module
-    :param conv:  'kaiming' or 'xavier'
-    :param batchnorm: 'normal' or 'constant'
-    :param linear: 'kaiming' or 'xavier'
-    :param lstm: 'kaiming' or 'xavier'
-    """
-    for m in model.modules():
-        if isinstance(m, (nn.modules.conv._ConvNd)):
-            if conv == 'kaiming':
-                initer.kaiming_normal_(m.weight)
-            elif conv == 'xavier':
-                initer.xavier_normal_(m.weight)
-            else:
-                raise ValueError("init type of conv error.\n")
-            if m.bias is not None:
-                initer.constant_(m.bias, 0)
-
-        elif isinstance(m, (nn.modules.batchnorm._BatchNorm)):
-            if batchnorm == 'normal':
-                initer.normal_(m.weight, 1.0, 0.02)
-            elif batchnorm == 'constant':
-                initer.constant_(m.weight, 1.0)
-            else:
-                raise ValueError("init type of batchnorm error.\n")
-            initer.constant_(m.bias, 0.0)
-
-        elif isinstance(m, nn.Linear):
-            if linear == 'kaiming':
-                initer.kaiming_normal_(m.weight)
-            elif linear == 'xavier':
-                initer.xavier_normal_(m.weight)
-            else:
-                raise ValueError("init type of linear error.\n")
-            if m.bias is not None:
-                initer.constant_(m.bias, 0)
-
-        elif isinstance(m, nn.LSTM):
-            for name, param in m.named_parameters():
-                if 'weight' in name:
-                    if lstm == 'kaiming':
-                        initer.kaiming_normal_(param)
-                    elif lstm == 'xavier':
-                        initer.xavier_normal_(param)
-                    else:
-                        raise ValueError("init type of lstm error.\n")
-                elif 'bias' in name:
-                    initer.constant_(param, 0)
-
-
-def group_weight(weight_group, module, lr):
-    group_decay = []
-    group_no_decay = []
-    for m in module.modules():
-        if isinstance(m, nn.Linear):
-            group_decay.append(m.weight)
-            if m.bias is not None:
-                group_no_decay.append(m.bias)
-        elif isinstance(m, nn.modules.conv._ConvNd):
-            group_decay.append(m.weight)
-            if m.bias is not None:
-                group_no_decay.append(m.bias)
-        elif isinstance(m, nn.modules.batchnorm._BatchNorm):
-            if m.weight is not None:
-                group_no_decay.append(m.weight)
-            if m.bias is not None:
-                group_no_decay.append(m.bias)
-    assert len(list(module.parameters())) == len(group_decay) + len(group_no_decay)
-    weight_group.append(dict(params=group_decay, lr=lr))
-    weight_group.append(dict(params=group_no_decay, weight_decay=.0, lr=lr))
-    return weight_group
-
-
-def colorize(gray, palette):
-    # gray: numpy array of the label and 1*3N size list palette
-    color = Image.fromarray(gray.astype(np.uint8)).convert('P')
-    color.putpalette(palette)
-    return color
-
-
-def find_free_port():
-    import socket
-    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    # Binding to port 0 will cause the OS to find an available port for us
-    sock.bind(("", 0))
-    port = sock.getsockname()[1]
-    sock.close()
-    # NOTE: there is still a chance the port could be taken by other processes.
-    return port
